Update part3a.md: change ids of the notes to strings in the code block
In the code block where hardcoded json values of notes are shown, the ids are integers; however, judging by the pictures on the page after that, the values are expected to have ids as strings. Moreover, the code snippet in the section "Fetching a single resource" does not work with ids as integers (as they are now); after changing the ids to strings it works fine.